Move all helpers into helpers_test.go
